+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## [0.32.0] - 2026-05-01
 
 ### Added
 
@@ -384,6 +384,7 @@ away in a future Tinty release.
 
 - Initial release
 
+[0.32.0]: https://github.com/tinted-theming/tinty/compare/v0.31.0...v0.32.0
 [0.31.0]: https://github.com/tinted-theming/tinty/compare/v0.30.0...v0.31.0
 [0.30.0]: https://github.com/tinted-theming/tinty/compare/v0.29.0...v0.30.0
 [0.29.0]: https://github.com/tinted-theming/tinty/compare/v0.28.0...v0.29.0